Added mole as image, and few changes in the code
1 - Replaced mole with an image:

Added a QPixmap for the mole image (mole.png), and set this image as an icon for the mole buttons.
Used setIcon() to display the mole image instead of the text "mole".
Ensured that the icon size matches the button size with setIconSize().

2 few changes in thew code:

added constants for easier code manipulation in the future